requirements.txt
fastapi[standard]
sqlmodel
pyodbc
requests
python-jose